Demonstrate locale resolution bug
Here I've added string resources for fr, fr-rCA, and pt-rPT. I've also created previews for each locale.

If you take a look at the previews, however, you'll see that the fr-rCA preview resolves to the fr value.

From what I can tell this is an issue with all subregions. The pt-rPT preview resolves correctly as of now, but if I added a pt resource, it would resolve incorrectly to the pt value instead.

I'd like to note, this also happens if you set up screenshot tests using @PreviewTest or Paparazzi.